Module Kv_lite__.Impl_with_lwt

include Kv_lite.Impl_intf.S_LWT

module M : sig ... end with type 'a t = 'a Lwt.t

type op = string * [ `Insert of string | `Delete ]

Operations are insert: (k,`Insert v), or delete: (k,`Delete)

type 'a or_error = ('a, string) Stdlib.result

val error_to_exn : 'a or_error -> 'a M.t

This unsafe function allows you to discard the error case; an exception will be raised at runtime; only use for testing!

type t

runtime db connection; error hook; error flag

val open_ : fn:string -> t or_error M.t

Error string if the db fails to open for some reason; kv-lite expects to use a lone database over which it has complete control; the default table name is "kv_lite" just in case you need it

val create : fn:string -> t or_error M.t

val close : t -> unit M.t

disconnect cannot throw an error in caqti

val set_error_hook : t -> (string -> unit) -> unit

Hook to catch post-initialization pre-close errors, which we hope do not occur; this allows the main insert/delete operations to avoid returning a result type. By default, this will exit the process, so be sure to set this in production code.

val slow_insert : t -> string -> string -> unit M.t

slow_insert t k v

val slow_delete : t -> string -> unit M.t

slow_delete t k v

val find_opt : t -> string -> string option M.t

find_opt t k

val batch : t -> op list -> unit M.t

following is fast because it batches multiple operations into a single transaction

val clear : t -> unit M.t

val last_batch_duration : t -> float

for profiling

val set_time : t -> (unit -> float) -> unit

update the standard timer (Sys.time) with a possibly more accurate one